requirements/dev.txt
django==2.0.3
psycopg2==2.7.4
pillow==5.0.0
pygraphviz==1.3.1
